Refactor code structure within pages>Results>pages
