+++ src/pages/Results/pages/Layout/components/EventStageBanner.tsx
